+++ Marlin/Marlin/Networking/MSIRouter.swift
@@ -23,7 +23,6 @@ enum MSIRouter: URLRequestConvertible
     case readNoticeToMariners(noticeNumber: Int64? = nil)
     
 //    static let baseURLString = "https://msi.nga.mil/api"
-//    static let ntmGraphicKeyBase = "16694429/SFH00000/UNTM"
     static let baseURLString = "https://msi.om.east.paas.nga.mil/api"
     static let ntmGraphicKeyBase = "16920957/SFH00000/UNTM"